Fix prop types
