+++ src/components/SlateEditor/plugins/SlateToolbar/SlateToolbar.jsx
@@ -13,7 +13,6 @@ import BEMHelper from 'react-bem-helper';
 import Types from 'slate-prop-types';
 import { Portal } from '../../../Portal';
 import ToolbarButton from './ToolbarButton';
-import { setActiveNode } from '../../createSlateStore';
 import { hasNodeOfType, checkSelectionForType } from '../../utils';
 import { TYPE as footnote } from '../footnote';
 import { TYPE as link } from '../link';
